+++ inst/include/quad/quad.h
@@ -1,21 +1,24 @@
 #ifndef QUAD_H
 #define QUAD_H
 
-#include <Rcpp.h>
+#include "cpp11.hpp"
+#include "cpp11/matrix.hpp"
+#include "cpp11/doubles.hpp"
+#include "cpp11/integers.hpp"
+
+using namespace cpp11;
+namespace writable = cpp11::writable;
 
 namespace quad {
- using namespace Rcpp;
 
-//IntegerVector quad_ib(IntegerVector nx, IntegerVector ny, LogicalVector ydown);
-//NumericVector quad_vb(IntegerVector nx, IntegerVector ny, LogicalVector ydown);
-
-inline IntegerVector quad_ib(IntegerVector nx, IntegerVector ny, LogicalVector ydown) {
+[[cpp11::register]]
+inline integers quad_ib(integers nx, integers ny, logicals ydown) {
     int len;
     int nc1;
     len = (nx[0]) * (ny[0]) * 4;
     nc1 = nx[0] + 1L; // nx columns
     //int nr1 = ny[0] + 1; // ny rows
-    IntegerVector quad(len);
+    writable::integers quad(len);
     int bl, br, tr, tl; // bottom left, bottom right, top left, top right
     int count = 0;
     for (int i = 0; i < nx[0]; i++) {
@@ -43,9 +46,9 @@ inline IntegerVector quad_ib(IntegerVector nx, IntegerVector ny, LogicalVector y
    return quad;
  }
 // flesh out x and y coordinates at edges of 2D array of nx*ny quads
- inline NumericVector quad_vb(IntegerVector nx, IntegerVector ny,
-                              LogicalVector ydown,
-                              LogicalVector zh) {
+ inline doubles quad_vb(integers nx, integers ny,
+                              logicals ydown,
+                              logicals zh) {
      int nc1 = nx[0] + 1;  // column edges
      int nr1 = ny[0] + 1;  // row edges
      int nrow = 2;
@@ -55,7 +58,7 @@ inline IntegerVector quad_ib(IntegerVector nx, IntegerVector ny, LogicalVector y
      double len = nc1 * nr1 * nrow; // total length of output
      int count = 0;   // ease chunking through loop
      // output vector
-     NumericVector vertex(len);
+     writable::doubles vertex(len);
 
      // step size from 0 to 1
      double dx = 1/(double)nx[0];
@@ -67,25 +70,28 @@ inline IntegerVector quad_ib(IntegerVector nx, IntegerVector ny, LogicalVector y
        dy = -dy;
      }
      // the 1D arrays across each margin, we calculate them first then expand-xy
-     NumericVector xx(nc1);
-     NumericVector yy(nr1);
+     writable::doubles xx(nc1);
+     writable::doubles yy(nr1);
      // populate x
      int i;
      for (i = 0; i < nc1; i++) {
         xx[i] = i * dx;
+      // Rprintf("%f\n", i * dx);
      }
     //  populate y
      for (int j = 0; j < nr1; j++) {
         yy[j] = ystart + j * dy;
+       //Rprintf("%f\n", j * dy);
+
      }
      for (int jj = 0; jj < nr1; jj++) {
          // increment columns first
          for (int ii = 0; ii < nc1; ii++) {
-           vertex[count + 0] = xx[ii];
-           vertex[count + 1] = yy[jj];
+           vertex[count + 0] = ii * dx; //(double)xx[ii];
+           vertex[count + 1] = ystart + jj * dy; //(double)yy[jj];
            if (zh[0]) {
-             vertex[count + 2] = 0; // z constant
-             vertex[count + 3] = 1; // h constant
+             vertex[count + 2] = 0.0; // z constant
+             vertex[count + 3] = 1.0; // h constant
 
            }
            count = count + nrow;
@@ -94,19 +100,25 @@ inline IntegerVector quad_ib(IntegerVector nx, IntegerVector ny, LogicalVector y
    return vertex;
  }
 // return matrix form of quad_vb (not interleaved) and optionally include z+h
-inline NumericMatrix quad_matrix_vb(IntegerVector nx, IntegerVector ny,
-                                    LogicalVector ydown,
-                                    LogicalVector zh) {
-  NumericVector v = quad_vb(nx, ny, ydown, zh);
+inline writable::doubles_matrix<>  quad_matrix_vb(integers nx, integers ny,
+                                    logicals ydown,
+                                    logicals zh) {
+  writable::doubles v = quad_vb(nx, ny, ydown, zh);
   int nr = 2;
   if (zh[0]) {
     nr = 4; // including z and h
   }
-  // Set number of rows and columns to attribute dim
-  v.attr("dim") = Dimension(nr, (nx[0] + 1) * (ny[0] + 1));
 
-  // Converting to Rcpp Matrix type
-  NumericMatrix quadvb_matrix = as<NumericMatrix>(v);
+  int nc = (nx[0] + 1) * (ny[0] + 1);
+  writable::doubles_matrix<> quadvb_matrix(nr, nc);
+  int cnt = 0;
+  for (int i = 0; i < nc; i++) {
+    for (int j = 0; j < nr; j++) {
+      // note j, i whereas with Rcpp we just set dim on the flat v and converted to matrix
+      quadvb_matrix(j, i) = (double)v[cnt];
+      cnt++;
+   }
+  }
   return quadvb_matrix;
 }
 }