+++ NEWS.md
@@ -1,3 +1,11 @@
+# textures dev
+
+* Convert to cpp11. 
+
+* Now using 'main' as default branch in git. 
+
+* Added 'extent' to `quad()`. 
+
 # textures 0.0.0.9020
 
 * quad.h now provides matrix version of vertex table, with or without z-h.